refactor(user): Extract profile cache to its own module
